crewai_guard/pyproject.toml
[project]
name = "crewai-guardrails"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "guardrails-ai>=0.6.3",
    "guardrails-api-client>=0.4.0a1",
    "crewai>=0.108.0",
    "crewai-tools>0.38.1",
]
